backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
packaging==25.0
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pytesseract==0.3.13
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
